+++ R/function_MapRegionalSources.R
@@ -304,11 +304,10 @@ MapRegionalSources <- function(data, map, map.subid.column = 1, group.column = N
 
       # Add various basemaps
       leafmap <- leafmap %>%
-        leaflet::addProviderTiles("CartoDB.Positron", group = "Map") %>%
+        leaflet::addProviderTiles("Esri.WorldGrayCanvas", group = "Map") %>%
         leaflet::addTiles(group = "Street") %>%
         leaflet::addProviderTiles("Esri.WorldTopoMap", group = "Topo") %>%
-        leaflet::addProviderTiles("Esri.WorldImagery", group = "Satellite") %>%
-        leaflet::addProviderTiles("CartoDB.PositronOnlyLabels", group = "Satellite")
+        leaflet::addProviderTiles("Esri.WorldImagery", group = "Satellite")
 
       # Save Image
       if (!file == "") {