+++ NEWS.md
@@ -1,5 +1,19 @@
 # HYPEtools NEWS
 
+## HYPEtools 1.6.8 (2026-09-02)
+*Enhancements and bug fixes*
+
+#### Highlights
+* Add ReadBestsims() function
+* Fix behavior of colors and drop in PlotPerformanceByAttribute()
+* Add additional geometry types to PlotParvalues()
+* Bug fixes to OutletNearObs()
+* Fix bug in na.color for PlotMapOutput()
+* Add CalculateRiverDistance() and LakeDataOutlets() functions
+* Add example [LakeData.txt](./inst/demo_model/LakeData.txt) file
+* Update examples for PlotSubbasinRouting() and fix bugs with line.weight and line.color
+* Replace CartoDB sources for Leaflet basemaps
+
 ## HYPEtools 1.6.7 (2026-04-08)
 *Enhancements and bug fixes*